+++ cucumber-android-hilt/src/test/java/io/cucumber/android/hilt/SomeCucumberHook.kt
@@ -1,0 +1,9 @@
+package io.cucumber.android.hilt
+
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class SomeCucumberHook @Inject constructor() {
+
+}