+++ cucumber-android-hilt/src/test/java/io/cucumber/android/hilt/StepsWithBaseClass.kt
@@ -1,0 +1,19 @@
+package io.cucumber.android.hilt
+
+import dagger.hilt.android.testing.HiltAndroidTest
+import javax.inject.Inject
+
+abstract class BaseSteps {
+
+    @Inject
+    lateinit var someCucumberHook: SomeCucumberHook
+
+    @Inject
+    lateinit var someDependencies: SomeDependencies
+}
+
+@HiltAndroidTest
+class StepsWithBaseClass:BaseSteps() {
+    @Inject
+    lateinit var someSingletonDependency: SomeSingletonDependency
+}