+++ cukeulator/src/androidTest/java/cucumber/cukeulator/test/ComposeRuleHolder.kt
@@ -4,16 +4,20 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import io.cucumber.junit.WithJunitRule
 import org.junit.Rule
+import javax.inject.Inject
+import javax.inject.Singleton
 
 @WithJunitRule("not @CustomComposable")
-class ComposeRuleHolder {
+@Singleton
+class ComposeRuleHolder @Inject constructor() {
 
     @get:Rule(order = 1)
     val composeRule = createEmptyComposeRule()
 }
 
 @WithJunitRule("@CustomComposable")
-class CustomComposableRuleHolder {
+@Singleton
+class CustomComposableRuleHolder @Inject constructor(){
 
     @get:Rule(order = 1)
     val composeRule = createComposeRule()