+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In Git
 
 * handle multiple classes (features) and methods (scenarios) specified in `class` argument to better align with tools which requests specific scenarios to be executed
+* support fields injection in steps classes with Hilt
 
 ## [7.14.0] - 2023-09-25